moved LoginDto and RegisterDto into an auth package inside dto package
