+++ api/src/main/java/com/chatop/api/config/SecurityConfig.java
@@ -14,6 +14,8 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 
+import com.chatop.api.service.AuthUserDetailsService;
+
 import lombok.RequiredArgsConstructor;
 
 @Configuration
