+++ api/src/main/java/com/chatop/api/dto/auth/RegisterDto.java
@@ -1,4 +1,4 @@
-package com.chatop.api.dto;
+package com.chatop.api.dto.auth;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
