+++ api/src/main/java/com/chatop/api/service/UserService.java
@@ -7,8 +7,8 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
-import com.chatop.api.dto.LoginDto;
-import com.chatop.api.dto.RegisterDto;
+import com.chatop.api.dto.auth.LoginDto;
+import com.chatop.api.dto.auth.RegisterDto;
 import com.chatop.api.model.User;
 import com.chatop.api.repository.UserRepository;
